+++ src/ui/buttons/props.js
@@ -473,6 +473,7 @@ export type RenderButtonProps = {|
   buttonSessionID: string,
   nonce: string,
   enableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
+  disableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
   components: $ReadOnlyArray<$Values<typeof COMPONENTS>>,
   onShippingChange: ?OnShippingChange,
   onShippingAddressChange: ?OnShippingAddressChange,
@@ -675,6 +676,7 @@ export type ButtonPropsInputs = {
   shopperSessionId?: string,
   nonce: string,
   enableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
+  disableFunding?: $ReadOnlyArray<?$Values<typeof FUNDING>>,
   components: $ReadOnlyArray<$Values<typeof COMPONENTS>>,
   onShippingChange: ?Function,
   onShippingAddressChange: ?Function,
@@ -813,27 +815,6 @@ export function determineRandomButtonColor({
   };
 }
 
-export function hasInvalidScriptOptionsForFullRedesign({
-  fundingSource,
-}: {|
-  fundingSource?: ?$Values<typeof FUNDING>,
-|}): boolean {
-  const validFundingSourcesForRedesign = [
-    undefined,
-    FUNDING.PAYPAL,
-    FUNDING.VENMO,
-    FUNDING.PAYLATER,
-    FUNDING.CREDIT,
-    FUNDING.CARD,
-  ];
-
-  if (validFundingSourcesForRedesign.includes(fundingSource)) {
-    return false;
-  }
-
-  return true;
-}
-
 export function throwErrorForInvalidButtonColor({
   fundingSource,
   fundingSourceColors,
@@ -918,13 +899,10 @@ export function getColorForFullRedesign({
     [BUTTON_COLOR.BLUE]: BUTTON_COLOR.REBRAND_BLUE,
     [BUTTON_COLOR.DARKBLUE]: BUTTON_COLOR.REBRAND_BLUE,
     [BUTTON_COLOR.GOLD]: BUTTON_COLOR.REBRAND_BLUE,
-
-    // not mapped yet since the styles are not setup
-    // These should never be hit since legacy experience should be set
     [BUTTON_COLOR.BLACK]: BUTTON_COLOR.REBRAND_BLACK,
     [BUTTON_COLOR.WHITE]: BUTTON_COLOR.REBRAND_WHITE,
     [BUTTON_COLOR.SILVER]: BUTTON_COLOR.REBRAND_WHITE,
-    [BUTTON_COLOR.DEFAULT]: BUTTON_COLOR.REBRAND_BLUE,
+    [BUTTON_COLOR.DEFAULT]: BUTTON_COLOR.DEFAULT,
 
     // normalizeButtonStyle gets called multiple times and
     // it can be called after color is already be mapped to rebranded style
@@ -964,9 +942,6 @@ export function getButtonColorExperience({
 }: GetButtonColorExperienceArgs): "abTest" | "fullRebrand" | "legacy" {
   const { isPaypalRebrandEnabled, isPaypalRebrandABTestEnabled } =
     experiment || {};
-  const rejectRedesign = hasInvalidScriptOptionsForFullRedesign({
-    fundingSource,
-  });
 
   if (!isPaypalRebrandEnabled) {
     return "legacy";
@@ -977,7 +952,7 @@ export function getButtonColorExperience({
     return fundingSource === FUNDING.PAYPAL ? "abTest" : "legacy";
   }
 
-  return rejectRedesign ? "legacy" : "fullRebrand";
+  return "fullRebrand";
 }
 
 export function getButtonColor({
@@ -1372,6 +1347,7 @@ export function normalizeButtonProps(
     sessionID = uniqueID(),
     buttonSessionID = uniqueID(),
     enableFunding,
+    disableFunding,
     components = [COMPONENTS.BUTTONS],
     nonce,
     onShippingChange,
@@ -1441,6 +1417,7 @@ export function normalizeButtonProps(
         fundingSource,
         fundingEligibility,
         enableFunding,
+        disableFunding,
         experiment,
         components,
         onShippingChange,
@@ -1485,6 +1462,7 @@ export function normalizeButtonProps(
     sessionID,
     nonce,
     enableFunding,
+    disableFunding,
     components,
     onShippingChange,
     onShippingAddressChange,